refactor: remove duplicate useTableSort hook
Fixes Issue #1 (CRITICAL) from CODE_REVIEW_FINDINGS.md

Changes:
- Deleted src/hooks/useTableSort.ts (hardcoded for GazaSite)
- Enhanced src/hooks/useTableSort.tsx with backwards compatibility
- Added 'sortedSites' alias to 'sortedData' for existing consumers
- Added SortField type export for backwards compatibility
- Updated imports in SitesTableDesktop and SitesTableMobile to use .tsx
- Updated CODE_REVIEW_FINDINGS.md progress tracker (1/20 complete, 5%)

Benefits:
- Single source of truth (DRY principle)
- Generic implementation with TypeScript generics
- Better performance with useCallback on handleSort
- Includes SortIcon component (previously duplicated in consumers)

Tests: All 1284 tests passing ✅
Lint: Clean ✅
