+++ CODE_REVIEW_FINDINGS.md
@@ -5,11 +5,14 @@
 **Review Date:** November 12, 2025
 **Codebase Size:** ~150+ source files (excluding tests)
 **Total Issues Found:** 20
+**Issues Resolved:** 1/20 (5%)
 **Severity Breakdown:**
-- **Critical:** 2 issues
-- **High:** 5 issues
-- **Medium:** 8 issues
-- **Low:** 5 issues
+- **Critical:** 2 issues (1 resolved ✅)
+- **High:** 5 issues (0 resolved)
+- **Medium:** 8 issues (0 resolved)
+- **Low:** 5 issues (0 resolved)
+
+**Progress:** 🟩⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜⬜ 5%
 
 ---
 
@@ -383,7 +386,7 @@
 ## Tracking Checklist
 
 ### Critical Issues
-- [ ] #1 - Remove duplicate `useTableSort.ts` hook
+- [x] #1 - Remove duplicate `useTableSort.ts` hook ✅ **COMPLETE**
 
 ### High Priority
 - [ ] #2 - Consolidate MOCK_DELAY constants to `src/constants/api.ts`
